+++ packages/paraswap/src/Paraswap.ts
@@ -1,14 +1,26 @@
-import { type SwapActionParams, compressJson } from '@rabbitholegg/questdk'
+import {
+  type SwapActionParams,
+  type StakeActionParams,
+  compressJson,
+  type ActionType,
+} from '@rabbitholegg/questdk'
 import { type Address } from 'viem'
-import { CHAIN_ID_ARRAY } from './chain-ids.js'
+import { STAKE_CHAIN_ID_ARRAY, SWAP_CHAIN_ID_ARRAY } from './chain-ids.js'
 import {
   constructGetTokens,
   constructAxiosFetcher,
   constructGetSpender,
 } from '@paraswap/sdk'
-import { PARASWAP_ABI } from './abi.js'
+import { PARASWAP_STAKE_ABI, PARASWAP_SWAP_ABI } from './abi.js'
 import axios from 'axios'
-import { DEFAULT_TOKEN_LIST_URL } from './contract-addresses.js'
+import {
+  DEFAULT_STAKE_TOKEN_LIST,
+  DEFAULT_SWAP_TOKEN_LIST,
+  MAINNET_SEPSP1_ADDRESS,
+  MAINNET_SEPSP2_ADDRESS,
+  OPTIMISM_SEPSP1_ADDRESS,
+  OPTIMISM_SEPSP2_ADDRESS,
+} from './contract-addresses.js'
 const fetcher = constructAxiosFetcher(axios) // alternatively constructFetchFetcher
 // If you're implementing swap or mint, simply duplicate this function and change the name
 export const swap = async (swap: SwapActionParams) => {
@@ -29,7 +41,7 @@ export const swap = async (swap: SwapActionParams) => {
     chainId: chainId, // The chainId of the source chain
     to: to,
     input: {
-      $abiAbstract: PARASWAP_ABI,
+      $abiAbstract: PARASWAP_SWAP_ABI,
       $or: [
         {
           assets:
@@ -95,21 +107,86 @@ export const swap = async (swap: SwapActionParams) => {
             },
           },
         },
+        {
+          data: {
+            fromToken: tokenIn,
+            fromAmount: amountIn,
+            toToken: tokenOut,
+            toAmount: amountOut,
+          },
+        },
       ],
     },
   })
 }
 
+export const stake = async (stake: StakeActionParams) => {
+  const { chainId, tokenOne, amountOne, tokenTwo } = stake
+  if (tokenOne !== undefined && tokenTwo !== undefined) {
+    const addressArray =
+      chainId === 1 ? [MAINNET_SEPSP2_ADDRESS] : [OPTIMISM_SEPSP2_ADDRESS]
+    return compressJson({
+      chainId: chainId,
+      to: {
+        $or: addressArray, // If both tokens are defined it has to be sePSP2
+      },
+      input: {
+        $abi: PARASWAP_STAKE_ABI,
+        $or: [
+          {
+            pspAmount: amountOne,
+          },
+          {
+            _assetAmount: amountOne,
+          },
+        ],
+      },
+    })
+  } else {
+    return compressJson({
+      chainId: chainId,
+      to: {
+        $or: [
+          OPTIMISM_SEPSP2_ADDRESS,
+          MAINNET_SEPSP2_ADDRESS,
+          OPTIMISM_SEPSP1_ADDRESS,
+          MAINNET_SEPSP1_ADDRESS,
+        ],
+      },
+      input: {
+        $abi: PARASWAP_STAKE_ABI,
+        $or: [
+          {
+            pspAmount: amountOne,
+          },
+          {
+            _assetAmount: amountOne,
+          },
+        ],
+      },
+    })
+  }
+}
+
 export const getSupportedTokenAddresses = async (
   _chainId: number,
+  actionType?: ActionType,
 ): Promise<Address[]> => {
+  if (actionType === undefined) return []
+  if (actionType === 'stake') {
+    return DEFAULT_STAKE_TOKEN_LIST[_chainId] as Address[]
+  }
   const { getTokens } = constructGetTokens({ chainId: _chainId, fetcher })
   // Default list only valid for
   return getTokens
     ? (await getTokens()).map((token) => token.address as Address)
-    : (DEFAULT_TOKEN_LIST_URL[_chainId] as Address[])
+    : (DEFAULT_SWAP_TOKEN_LIST[_chainId] as Address[])
 }
-
-export const getSupportedChainIds = async () => {
-  return CHAIN_ID_ARRAY
+export const getSupportedChainIds = async (actionType?: ActionType) => {
+  if (actionType === 'stake') {
+    return STAKE_CHAIN_ID_ARRAY
+  } else if (actionType === 'swap') {
+    return SWAP_CHAIN_ID_ARRAY
+  }
+  return []
 }